seperated out train and sample from phi_four
