+++ normflow/config.py
@@ -1,0 +1,5 @@
+"""
+config.py
+
+Module to parse runcards
+"""
